refactor(bootstrap): extract shared host bootstrap extensions, fix D7 serializer bug
